+++ ConvertToPowerShellNoteBook.ps1
@@ -10,13 +10,13 @@ function ConvertTo-PowerShellNoteBook {
     <#################################################################################################
     Parsing section.
     #################################################################################################>
-    
-    try{$CommentRanges = [System.Management.Automation.PSParser]::Tokenize((Get-Content $InputFileName), [ref]$null).Where({$_.Type -eq 'Comment'}) |
+
+    try{$CommentRanges = [System.Management.Automation.PSParser]::Tokenize((Get-Content -raw $InputFileName), [ref]$null).Where({$_.Type -eq 'Comment'}) |
     Select-Object -Property Start, Length, Type, Content }
     Catch{"This is not a valid PowerShell file"}
-    
+
     $s = Get-Content -Raw ( Resolve-Path $InputFileName )
-    
+
     $BlocksWitGaps = @()
     $Previous = $null
     foreach($CommentBlock in $CommentRanges ) {
@@ -31,12 +31,12 @@ function ConvertTo-PowerShellNoteBook {
         GapText = IF($CommentBlock.Start-$Previous.StopOffset -gt 1){[string] $s.Substring($Previous.StopOffset, ($CommentBlock.Start-$Previous.StopOffset)).trimstart()}else {[string] ''};
         Content = $CommentBlock.Content
         }
-    
+
         $Previous=$BlockOffsets
         $BlocksWitGaps+=[pscustomobject] $BlockOffsets
     }
-    
-    
+
+
     $AllBlocks = @()
     $Previous = $null
     $AllBlocks = $CommentRanges
@@ -58,7 +58,7 @@ function ConvertTo-PowerShellNoteBook {
                 Content = $s.Substring(($CommentBlock.Start+$CommentBlock.Length), ($s.Length-($CommentBlock.Start+$CommentBlock.Length))).trimstart()}
             }
         }
-    
+
     foreach($GapBlock in $BlocksWitGaps ) {
         $GapOffsets=[ordered]@{
         Start = $GapBlock.PreviousStopOffset;
@@ -67,35 +67,35 @@ function ConvertTo-PowerShellNoteBook {
         Type = 'Gap';
         Content = $GapBlock.GapText.trimstart()
         }
-    
+
         $Previous=$GapOffsets
         $AllBlocks+=if($GapOffsets.Length -gt 0){[pscustomobject] $GapOffsets}
     }
-    
+
     <#################################################################################################
     Notebook creation section.
-    #################################################################################################>    
-    
+    #################################################################################################>
+
     New-PSNotebook -NoteBookName $OutputNotebookName {
-    
+
         foreach($Block in $AllBlocks | Sort-Object Start ) {
-    
-            
+
+
             switch ($Block.Type) {
                 'Comment' {$TextBlock = $s.Substring($Block.Start, $Block.Length) -replace ("\n", "   `n  ")
-    
+
                             if($TextBlock.Trim().length -gt 0){
                                 Add-NotebookMarkdown -markdown (-join $TextBlock)
                             }
                         }
                 'Gap'   {$GapBlock = $s.Substring($Block.Start, $Block.Length)
-    
+
                             if($GapBlock.Trim().length -gt 0){
                                 (Add-NotebookCode -code (-join $GapBlock.trimstart().trimend())) -replace ("\n", "   `n  ")
                             }
                         }
             }
-    
+
         }
     }
 }